added /verbose command, refactored bolt_app.py, exported helper functions to utils.py
