+++ apps/.gitignore
@@ -1,3 +1,4 @@
 __pycache__
 **/.env
 slackbot/files/*
+testing*
