+++ apps/slackbot/utils.py
@@ -17,3 +17,48 @@ def load_files(files: List[str]) -> List[Document]:
     
     print(documents)
     return documents
+
+def contains_verbose(query: str) -> bool:
+    '''looks for -verbose in the question and returns True or False \
+    includes -verbosex'''
+    return "-verbose" in query.lower()
+
+def contains_verbosex(query: str) -> bool:
+    '''looks for -verbosex in the question and returns True or False'''
+    return "-verbosex" in query.lower()
+
+def format_verbose(logger):
+    '''Formats the logger into readable string'''
+    log_strings = []
+    for log in logger:
+        reply = log["reply"]
+        if "thoughts" in reply:
+            # reply = json.loads(reply)
+            formatted_reply = f"""-- Step: {log["Step"]} -- \nThoughts: \n {reply["thoughts"]} """
+            
+            if "command" in reply: # add command if it exists
+              formatted_reply += f"""\nCommand: \n {reply["command"]}"""
+            
+            log_strings.append(formatted_reply)
+
+        else: # for final response
+            formatted_reply = f"""-- Step: {log["Step"]} -- \nFinal Response: \n {reply}"""
+            log_strings.append(formatted_reply)
+    
+    log_string =  "\n".join(log_strings)
+    return log_string
+
+def show_commands_only(logger):
+    '''Modified version of format_verbose that only shows commands'''
+    log_strings = []
+    for log in logger:
+        reply = log["reply"]
+        if "command" in reply:
+          # reply = json.loads(reply)
+          formatted_reply = f"""-- Step: {log["Step"]} -- \nCommand: \n {reply["command"]}"""
+          log_strings.append(formatted_reply)
+        # else: # for final response
+        #   formatted_reply = f"""-- Step: {log["Step"]} -- \nFinal Response: \n {reply}"""
+        #   log_strings.append(formatted_reply)
+    log_string =  "\n".join(log_strings)
+    return log_string
